+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.18.0 (TBD)
 
 - [BREAKING] Incremented MSRV to 1.89.
+- [BREAKING] Adds DSA (EdDSA25519) and ECDH (X25519) using Curve25519 ([#537](https://github.com/0xMiden/crypto/pull/537)).
 
 ## 0.17.0 (2025-09-12)